+++ gpiozero/input_devices.py
@@ -4,6 +4,7 @@
 # Copyright (c) 2016-2019 Andrew Scheller <github@loowis.durge.org>
 # Copyright (c) 2015-2019 Dave Jones <dave@waveform.org.uk>
 # Copyright (c) 2015-2019 Ben Nuttall <ben@bennuttall.com>
+# Copyright (c) 2019 Kosovan Sofiia <sofiia.kosovan@gmail.com>
 # Copyright (c) 2018 Philippe Muller <philippe.muller@gmail.com>
 # Copyright (c) 2016 Steveis <SteveAmor@users.noreply.github.com>
 #